Add Music Pro custom widget to REX launcher
Implements a non-AppWidget custom widget system for Music Pro widget that:
- Appears in widget picker under "REX (debug)" category
- Displays currently playing song title, artist, and play/pause button
- Uses MediaSessionManager for real-time music metadata
- Updates periodically (500ms) for responsive UI
- Handles null safety (no crashes if no music playing)
- Supports drag/drop placement on workspace like standard widgets

New Files:
- MusicWidgetView.java: Custom FrameLayout showing music controls
- MusicWidgetProvider.java: Widget metadata and provider info
- CustomWidgetHostView.java: Container view for custom Java view widgets

Modified Files:
- CustomWidgetManager.java: Registers built-in custom widgets at init
- LauncherWidgetHolder.java: Detects and wraps Music Pro with CustomWidgetHostView
- widget_sections.xml: Added "REX (debug)" category with Music Pro registration
- strings.xml: Added category and widget strings

Architecture:
- Music Pro registers via CustomWidgetManager.registerBuiltInCustomWidgets()
- Avoids AppWidgetProvider and RemoteViews dependencies
- Widget placement hooks into standard LauncherWidgetHolder flow
- Graceful media session permission handling

Grid: 4x2 spans, lightweight performance
Stability: No null pointer risks, safe exception handling

+++ src/com/android/launcher3/PagedView.java
@@ -1835,7 +1835,7 @@ public abstract class PagedView<T extends View & PageIndicator> extends ViewGrou
         f *= 1.2f * Math.PI / 2.0f;
 
         float sineVal = (float) Math.sin(f);
-        float t = (f + Math.PI / 2.0f) / Math.PI; // normalize to 0-1
+        float t = (float) ((f + Math.PI / 2.0f) / Math.PI); // normalize to 0-1
         float eased = 1f - ((1f - t) * (1f - t) * (1f - t)); // cubic ease-out
 
         return Math.max(-1f, Math.min(1f, sineVal * (1f + 0.3f * eased)));
